// packages/api-types/src/enums.ts
export enum ResourceStatus {
  PROVISIONING = 'PROVISIONING',
  INACTIVE = 'INACTIVE',
  QUEUED = 'QUEUED',
  ACTIVE = 'ACTIVE',
  UPDATING = 'UPDATING',
  FAILED = 'FAILED',
  TERMINATING = 'TERMINATING',
  TERMINATED = 'TERMINATED',
  DELETING = 'DELETING',
  DELETED = 'DELETED',
}

export enum PortalType {
  CLOUDFLARE = 'CLOUDFLARE',
  DYNU = 'DYNU',
}

export enum TransponderMode {
  PROXY = 'PROXY',
}

export enum EventTypeKey {
  SYSTEM_RESET = 'SYSTEM_RESET',
  SYSTEM_RESET_FAILED = 'SYSTEM_RESET_FAILED',
  SYSTEM_RESET_SUCCESS = 'SYSTEM_RESET_SUCCESS',

  SYSTEM_RESET_HARD = 'SYSTEM_RESET_HARD',
  SYSTEM_RESET_HARD_FAILED = 'SYSTEM_RESET_HARD_FAILED',
  SYSTEM_RESET_HARD_SUCCESS = 'SYSTEM_RESET_HARD_SUCCESS',

  WORKER_UPDATE_SSH_KEYS = 'WORKER_UPDATE_SSH_KEYS',
  WORKER_UPDATE_SSH_KEYS_FAILED = 'WORKER_UPDATE_SSH_KEYS_FAILED',
  WORKER_SSH_KEYS_UPDATED = 'WORKER_SSH_KEYS_UPDATED',

  SYSTEM_TEST_EVENT = 'SYSTEM_TEST_EVENT',

  WORKER_CREATE = 'WORKER_CREATE',
  WORKER_CREATE_FAILED = 'WORKER_CREATE_FAILED',
  WORKER_CREATED = 'WORKER_CREATED',

  WORKER_UPDATE = 'WORKER_UPDATE',
  WORKER_UPDATED = 'WORKER_UPDATED',

  WORKER_START = 'WORKER_START',
  WORKER_STARTED = 'WORKER_STARTED',

  WORKER_TERMINATE = 'WORKER_TERMINATE',
  WORKER_TERMINATED = 'WORKER_TERMINATED',

  WORKER_DELETE = 'WORKER_DELETE',
  WORKER_DELETE_FAILED = 'WORKER_DELETE_FAILED',
  WORKER_DELETED = 'WORKER_DELETED',

  WORKER_IMAGE_CREATE = 'WORKER_IMAGE_CREATE',
  WORKER_IMAGE_CREATE_FAILED = 'WORKER_IMAGE_CREATE_FAILED',
  WORKER_IMAGE_CREATED = 'WORKER_IMAGE_CREATED',

  WORKER_DISK_CREATE = 'WORKER_DISK_CREATE',
  WORKER_DISK_CREATE_FAILED = 'WORKER_DISK_CREATE_FAILED',
  WORKER_DISK_CREATED = 'WORKER_DISK_CREATED',

  WORKER_DISK_ATTACH = 'WORKER_DISK_ATTACH',
  WORKER_DISK_ATTACH_FAILED = 'WORKER_DISK_ATTACH_FAILED',
  WORKER_DISK_ATTACHED = 'WORKER_DISK_ATTACHED',

  WORKER_DISK_DETACH = 'WORKER_DISK_DETACH',
  WORKER_DISK_DETACH_FAILED = 'WORKER_DISK_DETACH_FAILED',
  WORKER_DISK_DETACHED = 'WORKER_DISK_DETACHED',

  WORKER_DISK_DELETE = 'WORKER_DISK_DELETE',
  WORKER_DISK_DELETE_FAILED = 'WORKER_DISK_DELETE_FAILED',
  WORKER_DISK_DELETED = 'WORKER_DISK_DELETED',

  ATOM_CREATE = 'ATOM_CREATE',
  ATOM_CREATE_FAILED = 'ATOM_CREATE_FAILED',
  ATOM_CREATED = 'ATOM_CREATED',

  ATOM_START = 'ATOM_START',
  ATOM_START_FAILED = 'ATOM_START_FAILED',
  ATOM_STARTED = 'ATOM_STARTED',

  ATOM_TERMINATE = 'ATOM_TERMINATE',
  ATOM_TERMINATE_FAILED = 'ATOM_TERMINATE_FAILED',
  ATOM_TERMINATED = 'ATOM_TERMINATED',

  ATOM_DELETE = 'ATOM_DELETE',
  ATOM_DELETE_FAILED = 'ATOM_DELETE_FAILED',
  ATOM_DELETED = 'ATOM_DELETED',

  ZONE_CREATE = 'ZONE_CREATE',
  ZONE_CREATED = 'ZONE_CREATED',
  ZONE_DELETE = 'ZONE_DELETE',
  ZONE_DELETED = 'ZONE_DELETED',

  ZONE_START = 'ZONE_START',
  ZONE_START_FAILED = 'ZONE_START_FAILED',
  ZONE_STARTED = 'ZONE_STARTED',
  ZONE_STOP = 'ZONE_STOP',
  ZONE_STOP_FAILED = 'ZONE_STOP_FAILED',
  ZONE_STOPPED = 'ZONE_STOPPED',

  NODE_ASSIGN_WORKER = 'NODE_ASSIGN_WORKER',
  NODE_ASSIGNED_WORKER = 'NODE_ASSIGNED_WORKER',
  NODE_UNASSIGN_WORKER = 'NODE_UNASSIGN_WORKER',
  NODE_UNASSIGNED_WORKER = 'NODE_UNASSIGNED_WORKER',

  NODE_START = 'NODE_START',
  NODE_START_FAILED = 'NODE_START_FAILED',
  NODE_STARTED = 'NODE_STARTED',
  NODE_STOP = 'NODE_STOP',
  NODE_STOP_FAILED = 'NODE_STOP_FAILED',
  NODE_STOPPED = 'NODE_STOPPED',

  NODE_CREATE_FIBER = 'NODE_CREATE_FIBER',
  NODE_CREATE_FIBER_FAILED = 'NODE_CREATE_FIBER_FAILED',
  NODE_FIBER_CREATED = 'NODE_FIBER_CREATED',

  NODE_UPDATE_FIBER = 'NODE_UPDATE_FIBER',
  NODE_UPDATE_FIBER_FAILED = 'NODE_UPDATE_FIBER_FAILED',
  NODE_FIBER_UPDATED = 'NODE_FIBER_UPDATED',

  NODE_DELETE_FIBER = 'NODE_DELETE_FIBER',
  NODE_DELETE_FIBER_FAILED = 'NODE_DELETE_FIBER_FAILED',
  NODE_FIBER_DELETED = 'NODE_FIBER_DELETED',

  NODE_START_FIBER = 'NODE_START_FIBER',
  NODE_START_FIBER_FAILED = 'NODE_START_FIBER_FAILED',
  NODE_FIBER_STARTED = 'NODE_FIBER_STARTED',
  NODE_STOP_FIBER = 'NODE_STOP_FIBER',
  NODE_STOP_FIBER_FAILED = 'NODE_STOP_FIBER_FAILED',
  NODE_FIBER_STOPPED = 'NODE_FIBER_STOPPED',

  PORTAL_CREATE = 'PORTAL_CREATE',
  PORTAL_CREATE_FAILED = 'PORTAL_CREATE_FAILED',
  PORTAL_CREATED = 'PORTAL_CREATED',

  PORTAL_UPDATE = 'PORTAL_UPDATE',
  PORTAL_UPDATE_FAILED = 'PORTAL_UPDATE_FAILED',
  PORTAL_UPDATED = 'PORTAL_UPDATED',

  PORTAL_SYNC = 'PORTAL_SYNC',
  PORTAL_SYNC_FAILED = 'PORTAL_SYNC_FAILED',
  PORTAL_SYNCED = 'PORTAL_SYNCED',

  PORTAL_DELETE = 'PORTAL_DELETE',
  PORTAL_DELETE_FAILED = 'PORTAL_DELETE_FAILED',
  PORTAL_DELETED = 'PORTAL_DELETED',

  TRANSPONDER_CREATE = 'TRANSPONDER_CREATE',
  TRANSPONDER_CREATE_FAILED = 'TRANSPONDER_CREATE_FAILED',
  TRANSPONDER_CREATED = 'TRANSPONDER_CREATED',

  TRANSPONDER_UPDATE = 'TRANSPONDER_UPDATE',
  TRANSPONDER_UPDATE_FAILED = 'TRANSPONDER_UPDATE_FAILED',
  TRANSPONDER_UPDATED = 'TRANSPONDER_UPDATED',

  TRANSPONDER_DELETE = 'TRANSPONDER_DELETE',
  TRANSPONDER_DELETE_FAILED = 'TRANSPONDER_DELETE_FAILED',
  TRANSPONDER_DELETED = 'TRANSPONDER_DELETED',
}